AP-300 rename PaymentMethodService to PaymentMethodOptionsBuilder
